fix: fix CI build errors in hadris-udf-cli and hadris-fat exfat
- Replace copy-pasted hadris-iso-cli code in hadris-udf-cli with a
  working implementation that uses hadris-udf: implements info, ls,
  tree, and create commands
- Add hadris-udf-cli to workspace members in Cargo.toml
- Fix type mismatch in hadris-fat exfat fs.rs: cast size_of result
  to u64 when adding to u64 offset variable

+++ crates/hadris-fat/src/exfat/fs.rs
@@ -124,7 +124,7 @@ where
                     _ => {}
                 }
 
-                offset += size_of::<RawDirectoryEntry>();
+                offset += size_of::<RawDirectoryEntry>() as u64;
                 entries_read += 1;
             }
